Fix submitting form without address
